requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
Django==2.1.3
django-tables2==2.0.4
EasyProcess==0.2.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
parsel==1.5.1
pika==0.12.0
Pillow==5.3.0
pkg-resources==0.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytesseract==0.2.5
python-scrapyd-api==2.1.2
pytz==2018.7
PyVirtualDisplay==0.2.1
queuelib==1.5.0
requests==2.20.1
Scrapy==1.5.1
scrapyd==1.2.0
selenium==3.141.0
service-identity==17.0.0
six==1.11.0
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.6.0
